Add files via upload
This Python project involves web scraping Flipkart's mobile section to extract data on various mobile phones. The web scraping is done using the BeautifulSoup and requests libraries in Python. The code retrieves information such as the mobile phone name, price, and rating. This data is then stored in a CSV file for easy analysis and further processing.

The project demonstrates the ability to automate data collection from a website using Python, which can be used in various industries such as e-commerce, finance, and marketing. The extracted data can be analyzed to gain insights into market trends, consumer preferences, and competitive intelligence.

Overall, this project showcases the ability to apply Python programming and web scraping techniques to solve real-world problems and generate valuable insights.